pyproject.toml
[build-system]
requires = ["setuptools", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "api_dock"
version = "0.0.1"
description = "A flexible API gateway that allows you to proxy requests to multiple remote APIs and Databases"
readme = "README.md"
license = {text = "BSd 3-clause"}
authors = [
  {name = "Brookie Guzder-Williams", email = "bguzder-williams@berkeley.edu"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering",
]
requires-python = ">=3.8"
dependencies = [
    "click",
    "fastapi",
    "uvicorn",
    "pyyaml",
    "httpx",
    "pyarrow",
    "duckdb"
]

[project.scripts]
api-dock = "api_dock.cli:main"

[tool.setuptools]
packages = [
	"api_dock",
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-64", "osx-arm64"]

[tool.pixi.pypi-dependencies]
api_dock = { path = ".", editable = true }

[tool.pixi.dependencies]
pycodestyle = ">=2.14.0,<3"
ipykernel = ">=6.30.1,<7"
jupyterlab = ">=4.4.9,<5"
numpy = ">=2.3.3,<3"
pandas = ">=2.3.2,<3"
fastapi = ">=0.117.1,<0.118"
uvicorn = ">=0.37.0,<0.38"
pyyaml = ">=6.0.3,<7"
httpx = ">=0.28.1,<0.29"
flask = ">=3.1.2,<4"
duckdb = ">=1.1.3,<2"

[tool.pixi.environments]
default = { features = [] }
dev = { features = ["dev"] }

[tool.pixi.feature.dev.dependencies]
twine = "*"

[tool.pixi.feature.dev.pypi-dependencies]
build = "*"

